moved 'rotation' code into extension
